refactor: Create types file
